+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes to this project are documented in this file. More detail and links can b
 - AUTOTOOL-1169: [GitHub #9](https://github.com/F5Networks/f5-telemetry-streaming/issues/9): Add new Prometheus Pull Consumer
 - AUTOTOOL-1449: Add information about inlined pollers (System Poller and iHealthPoller) to TEEM report
 ### Fixed
+- AUTOTOOL-1506: Added a timeout to Azure metadata service HTTP requests
 ### Changed
 ### Removed